MainQuest/MainQuest1/requirements.txt
absl-py==0.12.0
aiohttp
aiosignal
albumentations==1.1.0
antlr4-python3-runtime==4.8
anyio
appdirs==1.4.4
argon2-cffi
astunparse==1.6.3
async-generator==1.10
async-timeout
attrs
audioread==2.1.9
backcall
backports.functools-lru-cache
beautifulsoup4==4.6.0
bleach
branca==0.4.2
brotlipy
cachetools==4.2.4
certifi==2021.10.8
cffi
chardet==3.0.4
charset-normalizer
clang==5.0
click==8.0.3
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.0.0
cmake==3.21.3
colorama
cryptography
cssselect==1.1.0
customized-konlpy==0.0.64
cycler==0.11.0
Cython==0.29.24
dataclasses==0.6
datasets==1.14.0
debugpy
decorator==4.4.2
defusedxml
delayed==0.11.0b1
Deprecated==1.2.13
dill==0.3.4
dlib==19.22.1
docutils==0.18.1
editdistance==0.6.0
efficientnet==1.0.0
efficientnet-pytorch==0.7.1
entrypoints
essential-generators==1.0
face-recognition==1.3.0
face-recognition-models==0.3.0
feedfinder2==0.0.4
feedparser==6.0.8
filelock
keras==2.6.0
Keras-Applications==1.0.8
keras-ocr==0.8.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
konlpy==0.5.2
labelme2coco==0.1.2
librosa==0.8.1
lightgbm==3.3.0
llvmlite==0.36.0
lmdb==1.2.1
loguru==0.5.3
lxml==4.6.3
matplotlib==3.4.3
matplotlib-venn==0.11.6
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
opencv-python-headless==4.5.4.60
pandas==1.3.3
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-addons==0.14.0
tensorflow-datasets==4.4.0
tensorflow-estimator==2.7.0
tokenizers==0.10.3
transformers==4.11.3
xgboost==1.4.2
xmljson==0.2.1